refactor: solved majority of comments in the PR review
